Zebraix: Mostly moving code into separate files.
