+++ zebraix/src/z-graph/src/jaywalk_graph/mod.rs
@@ -15,6 +15,12 @@
 pub mod jaywalk_foundation;
 pub mod jaywalk_traiting;
 pub mod zgraph_base;
+pub mod zgraph_builtins;
+pub mod zgraph_graphdef;
+pub mod zgraph_machine;
+pub mod zgraph_registry;
+pub mod zgraph_solver;
+pub mod zgraph_svg;
 
 // use crate::jaywalk_graph::jaywalk_foundation::EmptyVec;
 use crate::jaywalk_graph::jaywalk_traiting::absent_f64;
