+++ zebraix/src/z-graph/test-files/golden-inputs/simple_graph.json
@@ -1,6 +1,37 @@
 {
   "name": "A test graph",
   "description": "",
-  "nodes": [],
+  "nodes": [
+    {
+      "name": "input",
+      "node_type": {
+        "name": "gather"
+      }
+    },
+    {
+      "name": "test text",
+      "node_type": {
+        "name": "test_text"
+      }
+    },
+    {
+      "name": "text circle",
+      "node_type": {
+        "name": "test_circle"
+      }
+    },
+    {
+      "name": "main group",
+      "node_type": {
+        "name": "group"
+      }
+    },
+    {
+      "name": "output",
+      "node_type": {
+        "name": "gather"
+      }
+    }
+  ],
   "edges": []
 }
